requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
filelock==3.0.12
Flask==2.0.1
google-api-core==1.31.0
google-auth==1.33.1
google-cloud==0.34.0
google-cloud-appengine-logging==0.1.1
google-cloud-audit-log==0.1.0
google-cloud-core==1.7.1
google-cloud-logging==2.5.0
googleapis-common-protos==1.53.0
grpcio==1.38.1
gunicorn==20.1.0
huggingface-hub==0.0.12
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.21.1
packaging==21.0
pkg-resources==0.0.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
sacremoses==0.0.45
six==1.16.0
tokenizers==0.10.3
torch==1.9.0
tqdm==4.61.2
transformers==4.8.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
